requirements.txt
google-cloud-bigquery
google-genai
python-dotenv
pandas